[refactor] 상품 목록 페이지 필터버튼 active 로직 categoryId -> subCategoryId로 수정, 파일 위치 변경
